pyproject.toml
[project]
name = "aiidalab-qe-base"
description = "Utilities, widgets, and components shared by the AiiDAlab Quantum ESPRESSO app and its plugins"
authors = [
    { name = "AiiDAlab team", email = "aiidalab@materialscloud.org" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "aiida-core~=2.5,<3",
    "aiidalab-widgets-base~=2.4.0",
    "aiida-quantumespresso~=4.12.0",
    "anywidget==0.9.13",
    "table_widget~=0.0.2",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pre-commit~=3.2",
]
test = [
    "pytest~=7.4",
    "pytest-cov~=5.0",
]

[project.urls]
Homepage = "https://github.com/aiidalab/aiidalab-qe-base"
Repository = "https://github.com/aiidalab/aiidalab-qe-base"

[project.entry-points."aiida.workflows"]
"aiidalab_qe_base.testing.workflows.mock" = "aiidalab_qe_base.tests.conftest:MockWorkChain"

[build-system]
requires = ["flit_core~=3.4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "aiidalab_qe_base"

[tool.flit.sdist]
exclude = [
    'docs/',
    'tests/',
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning",
    "ignore:.*zmq.eventloop.ioloop is deprecated.*:DeprecationWarning",
    "ignore:.*use the instance .metadata dictionary directly.*:DeprecationWarning",
    "ignore:Object of type .* not in session, .* operation along .* will not proceed:sqlalchemy.exc.SAWarning"
]
